move document trait implementations into libs/places
