+++ libs/common/Cargo.toml
@@ -11,5 +11,4 @@ categories = [ "application" ]
 readme = "README.md"
 
 [dependencies]
-places = { path = "../places" }
 serde = { version = "1.0", default_features = false }
