+++ libs/places/Cargo.toml
@@ -19,3 +19,4 @@ cosmogony = "0.10"
 transit_model = "0.31.4"
 typed_index_collection = "1.1"
 navitia-poi-model = "0.3"
+common = { path = "../common" }
