+++ libs/places/src/street.rs
@@ -1,3 +1,4 @@
+use common::document::{ContainerDocument, Document};
 use geojson::Geometry;
 use serde::{Deserialize, Serialize};
 use std::sync::Arc;
@@ -47,6 +48,18 @@ impl Members for Street {
     }
     fn admins(&self) -> Vec<Arc<Admin>> {
         self.administrative_regions.clone()
+    }
+}
+
+impl Document for Street {
+    fn id(&self) -> String {
+        self.id.clone()
+    }
+}
+
+impl ContainerDocument for Street {
+    fn static_doc_type() -> &'static str {
+        "street"
     }
 }
